hal: moved target-specific hal_version code to a separate file
