+++ firmware/targets/f7/furi_hal/furi_hal_version.c
@@ -197,18 +197,6 @@ void furi_hal_version_init() {
     FURI_LOG_I(TAG, "Init OK");
 }
 
-bool furi_hal_version_do_i_belong_here() {
-    return (furi_hal_version_get_hw_target() == 7) || (furi_hal_version_get_hw_target() == 0);
-}
-
-const char* furi_hal_version_get_model_name() {
-    return "Flipper Zero";
-}
-
-const char* furi_hal_version_get_model_code() {
-    return "FZ.1";
-}
-
 FuriHalVersionOtpVersion furi_hal_version_get_otp_version() {
     if(*(uint64_t*)FURI_HAL_VERSION_OTP_ADDRESS == 0xFFFFFFFF) {
         return FuriHalVersionOtpVersionEmpty;
